+++ app/routes/explore.route.tsx
@@ -5,7 +5,8 @@ import { useEffect, useState } from "react";
 import recipePlaceholder from "~/assets/images/recipe-placeholder.jpg";
 import { InfiniteScroller } from "~/components/infinite-scroller";
 import Layout from "~/components/layout";
-import RecipeCard from "~/recipes/recipes/recipe-card";
+import { InvisibleTooltip } from "~/components/tooltip";
+import RecipeCard from "~/recipes/recipe-card";
 import { RecipesResponse, getRecipes } from "~/recipes/recipe.server";
 
 const DEFAULT_SKIP = 0;
@@ -70,13 +71,19 @@ export default function ExplorePage() {
               );
             })}
           </div>
-          <button
-            className="rounded bg-slate-600 px-4 py-2 text-blue-100 text-xl hover:bg-blue-500 active:bg-blue-600 disabled:bg-gray-400 disabled:cursor-not-allowed"
-            disabled={!recipesAvailable}
-            onClick={loadNext}
-          >
-            Load more
-          </button>
+          <div>
+            <InvisibleTooltip
+              displayMessage={!recipesAvailable}
+              message={"No more recipes to load!"}
+            >
+              <button
+                className="rounded bg-slate-600 px-4 py-2 text-blue-100 text-xl hover:bg-blue-500 active:bg-blue-600 disabled:bg-gray-400 disabled:cursor-not-allowed"
+                disabled={!recipesAvailable}
+              >
+                Load More
+              </button>
+            </InvisibleTooltip>
+          </div>
         </main>
       </Layout>
     </InfiniteScroller>
